Annotation Adjuster
Starting to write the python script to edit `.gff` information to match
with the annotation information inside of a `.sam` file.

Need to figure out how BWA aligns the nodes to the reference genome so
that I can properly adjust the annotations. Will also need to edit the
variables (lists) that I am creating (there might be a few too many).