Parsing Cigar Strings
Filtered which nodes were being parsed further by tag (2048 or 0x800).
Was able to get the sequences to be edited correctly (it looks that
way) by parsing the cigar string and modifying the `seq` accordingly.

Will need to make sure that annotations from `.gff` file are checked
and 'moved' while the sequences are edited so that their position
changes correctly.

Need to re-look into nodes (parts of the alignment [`.sam` file]) that
were removed just to make sure we get proper wrapping of the
annotations.

Will also need to clean up some of the test strings and comments to
ensure readability for later because who wants to come back a couple
years from now and have no idea what's going on...