Fixed Sequence Parsing
Sequences are now reversed where necessary before they are processed
by their cigar string.

Created a test `for` loop that further inflates the sequence with a
temp variable '$'. just realized that I need to replace the base at
`base_num-1` with the temp variable instead of inflating. This way
the sequence length is maintained instead of modified; annotations
can be mapped to the 'new' sequence, and modified when the '$' are
removed from the reads.

when working `.gff` output is confirmed the number of lists will also
be reduced.
  - I created some additional lists to parse the `.sam` data
      instead of modifying the original data
  - Once everything is working the original lists can be modified
      instead of new lists