Adjusted Features
It seems as though, right now, that I was able to correctly adjust
both the sequences and the annotations. Some of the annotations will
be duplicated to fit to the chimeric alignments that are present in the
`.sam` file.

All that remains to be done is to re-write the `.gff` with the new
locations for the features and the new sequences. Aftef the `.gff` file
is written, I will make a new `.sam` file that has adjusted names for
the repeated nodes (EX: NODE_3811_length_7200_cov_9 --> NODE_3811_p1
and NODE_3811_p2).

Nodes without repeats will still be named following the same
convention. EX: NODE_7_p1